+++ invoice_line_revenue_distribution_operating_unit/__init__.py
@@ -3,5 +3,4 @@
 # Copyright 2017 Willem Hulshof ((www.magnus.nl).)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-# from . import controllers
 from . import models
