--- invoice_line_revenue_distribution_operating_unit/controllers/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import controllers
