--- invoice_line_revenue_distribution_operating_unit/demo/demo.xml
@@ -1,30 +1,0 @@
-<odoo>
-    <data>
-        <!--  -->
-        <!--   <record id="object0" model="invoice_line_revenue_distribution_operating_unit.invoice_line_revenue_distribution_operating_unit"> -->
-        <!--     <field name="name">Object 0</field> -->
-        <!--     <field name="value">0</field> -->
-        <!--   </record> -->
-        <!--  -->
-        <!--   <record id="object1" model="invoice_line_revenue_distribution_operating_unit.invoice_line_revenue_distribution_operating_unit"> -->
-        <!--     <field name="name">Object 1</field> -->
-        <!--     <field name="value">10</field> -->
-        <!--   </record> -->
-        <!--  -->
-        <!--   <record id="object2" model="invoice_line_revenue_distribution_operating_unit.invoice_line_revenue_distribution_operating_unit"> -->
-        <!--     <field name="name">Object 2</field> -->
-        <!--     <field name="value">20</field> -->
-        <!--   </record> -->
-        <!--  -->
-        <!--   <record id="object3" model="invoice_line_revenue_distribution_operating_unit.invoice_line_revenue_distribution_operating_unit"> -->
-        <!--     <field name="name">Object 3</field> -->
-        <!--     <field name="value">30</field> -->
-        <!--   </record> -->
-        <!--  -->
-        <!--   <record id="object4" model="invoice_line_revenue_distribution_operating_unit.invoice_line_revenue_distribution_operating_unit"> -->
-        <!--     <field name="name">Object 4</field> -->
-        <!--     <field name="value">40</field> -->
-        <!--   </record> -->
-        <!--  -->
-    </data>
-</odoo>
