+++ invoice_line_revenue_distribution_operating_unit/models/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from . import models
+from . import analytic
+from . import invoice
+from . import sale
